Split etserver into etserver & etterminal (#123)
* split etserver into etserver/etterminal

* support -x with etterminal & etserver
