+++ src/terminal/TerminalServer.cpp
@@ -48,20 +48,10 @@ using namespace google;
 using namespace gflags;
 
 #define BUF_SIZE (16 * 1024)
-const int KEEP_ALIVE_DURATION = 7;
 
 DEFINE_int32(port, 0, "Port to listen on");
-DEFINE_string(idpasskey, "",
-              "If set, uses IPC to send a client id/key to the server daemon");
-DEFINE_string(idpasskeyfile, "",
-              "If set, uses IPC to send a client id/key to the server daemon "
-              "from a file");
 DEFINE_bool(daemon, false, "Daemonize the server");
 DEFINE_string(cfgfile, "", "Location of the config file");
-DEFINE_bool(jump, false,
-            "If set, forward all packets between client and dst terminal");
-DEFINE_string(dsthost, "", "Must be set if jump is set to true");
-DEFINE_int32(dstport, 2022, "Must be set if jump is set to true");
 DEFINE_int32(v, 0, "verbose level");
 DEFINE_bool(logtostdout, false, "log to stdout");
 
@@ -70,38 +60,6 @@ shared_ptr<UserTerminalRouter> terminalRouter;
 vector<shared_ptr<thread>> terminalThreads;
 mutex terminalThreadMutex;
 bool halt = false;
-string getIdpasskey() {
-  string idpasskey = FLAGS_idpasskey;
-  if (FLAGS_idpasskeyfile.length() > 0) {
-    // Check for passkey file
-    std::ifstream t(FLAGS_idpasskeyfile.c_str());
-    std::stringstream buffer;
-    buffer << t.rdbuf();
-    idpasskey = buffer.str();
-    // Trim whitespace
-    idpasskey.erase(idpasskey.find_last_not_of(" \n\r\t") + 1);
-    // Delete the file with the passkey
-    remove(FLAGS_idpasskeyfile.c_str());
-  }
-  return idpasskey;
-}
-
-void setDaemonLogFile(string idpasskey, string daemonType) {
-  string first_idpass_chars = idpasskey.substr(0, 10);
-  string std_file =
-      string("/tmp/etserver_") + daemonType + "_" + first_idpass_chars;
-#if __NetBSD__
-  FILE *stdout_stream = freopen("/tmp/etclient_err", "w+", stdout);
-  setvbuf(stdout_stream, NULL, _IOLBF, BUFSIZ);  // set to line buffering
-  FILE *stderr_stream = freopen("/tmp/etclient_err", "w+", stderr);
-  setvbuf(stderr_stream, NULL, _IOLBF, BUFSIZ);  // set to line buffering
-#else
-  stdout = fopen(std_file.c_str(), "w+");
-  setvbuf(stdout, NULL, _IOLBF, BUFSIZ);  // set to line buffering
-  stderr = fopen(std_file.c_str(), "w+");
-  setvbuf(stderr, NULL, _IOLBF, BUFSIZ);  // set to line buffering
-#endif
-}
 
 void runJumpHost(shared_ptr<ServerClientConnection> serverClientState) {
   bool run = true;
@@ -389,144 +347,6 @@ void startServer() {
   }
 }
 
-void startUserTerminal(string idpasskey) {
-  UserTerminalHandler uth;
-  uth.connectToRouter(idpasskey);
-  cout << "IDPASSKEY:" << idpasskey << endl;
-  if (::daemon(0, 0) == -1) {
-    LOG(FATAL) << "Error creating daemon: " << strerror(errno);
-  }
-  setDaemonLogFile(idpasskey, "terminal");
-  uth.run();
-}
-
-void startJumpHostClient(string idpasskey) {
-  cout << "IDPASSKEY:" << idpasskey << endl;
-  auto idpasskey_splited = split(idpasskey, '/');
-  string id = idpasskey_splited[0];
-  string passkey = idpasskey_splited[1];
-
-  string host = FLAGS_dsthost;
-  int port = FLAGS_dstport;
-
-  if (::daemon(0, 0) == -1) {
-    LOG(FATAL) << "Error creating daemon: " << strerror(errno);
-  }
-  setDaemonLogFile(idpasskey, "jumphost");
-  sockaddr_un remote;
-
-  int routerFd = ::socket(AF_UNIX, SOCK_STREAM, 0);
-  FATAL_FAIL(routerFd);
-  remote.sun_family = AF_UNIX;
-  strcpy(remote.sun_path, ROUTER_FIFO_NAME);
-
-  if (connect(routerFd, (struct sockaddr *)&remote, sizeof(sockaddr_un)) < 0) {
-    close(routerFd);
-    if (errno == ECONNREFUSED) {
-      cout << "Error:  The Eternal Terminal daemon is not running.  Please "
-              "(re)start the et daemon on the server."
-           << endl;
-    } else {
-      cout << "Error:  Connection error communicating with et deamon: "
-           << strerror(errno) << "." << endl;
-    }
-    exit(1);
-  }
-
-  try {
-    RawSocketUtils::writeMessage(routerFd, idpasskey);
-  } catch (const std::runtime_error &re) {
-    LOG(FATAL) << "Cannot send idpasskey to router: " << re.what();
-  }
-
-  InitialPayload payload;
-
-  shared_ptr<SocketHandler> jumpclientSocket(new UnixSocketHandler());
-  shared_ptr<ClientConnection> jumpclient = shared_ptr<ClientConnection>(
-      new ClientConnection(jumpclientSocket, host, port, id, passkey));
-
-  int connectFailCount = 0;
-  while (true) {
-    try {
-      jumpclient->connect();
-      jumpclient->writeProto(payload);
-    } catch (const runtime_error &err) {
-      LOG(ERROR) << "Connecting to dst server failed: " << err.what();
-      connectFailCount++;
-      if (connectFailCount == 3) {
-        LOG(INFO) << "Could not make initial connection to dst server";
-        cout << "Could not make initial connection to " << host << ": "
-             << err.what() << endl;
-        exit(1);
-      }
-      continue;
-    }
-    break;
-  }
-  VLOG(1) << "JumpClient created with id: " << jumpclient->getId();
-
-  bool run = true;
-  time_t keepaliveTime = time(NULL) + KEEP_ALIVE_DURATION;
-
-  while (run && !jumpclient->isShuttingDown()) {
-    // Data structures needed for select() and
-    // non-blocking I/O.
-    fd_set rfd;
-    timeval tv;
-
-    FD_ZERO(&rfd);
-    FD_SET(routerFd, &rfd);
-    int maxfd = routerFd;
-    int jumpClientFd = jumpclient->getSocketFd();
-    if (jumpClientFd > 0) {
-      FD_SET(jumpClientFd, &rfd);
-      maxfd = max(maxfd, jumpClientFd);
-    }
-    tv.tv_sec = 0;
-    tv.tv_usec = 10000;
-    select(maxfd + 1, &rfd, NULL, NULL, &tv);
-
-    try {
-      // forward local router -> DST terminal.
-      if (FD_ISSET(routerFd, &rfd)) {
-        keepaliveTime = time(NULL) + KEEP_ALIVE_DURATION;
-        if (jumpClientFd < 0) {
-          LOG(INFO) << "User comes back, reconnecting";
-          jumpclient->closeSocket();
-          sleep(3);
-          continue;
-        } else {
-          string s = RawSocketUtils::readMessage(routerFd);
-          jumpclient->writeMessage(s);
-          VLOG(3) << "Sent message from router to dst terminal: " << s.length();
-        }
-      }
-      // forward DST terminal -> local router
-      if (jumpClientFd > 0 && FD_ISSET(jumpClientFd, &rfd)) {
-        while (jumpclient->hasData()) {
-          string receivedMessage;
-          jumpclient->readMessage(&receivedMessage);
-          RawSocketUtils::writeMessage(routerFd, receivedMessage);
-          VLOG(3) << "Send message from dst terminal to router: "
-                  << receivedMessage.length();
-        }
-        keepaliveTime = time(NULL) + KEEP_ALIVE_DURATION;
-      }
-      // src disconnects, close jump -> dst
-      if (jumpClientFd > 0 && keepaliveTime < time(NULL)) {
-        LOG(INFO) << "Jumpclient idle, killing connection";
-        jumpclient->Connection::closeSocket();
-      }
-    } catch (const runtime_error &re) {
-      LOG(ERROR) << "Error: " << re.what();
-      cout << "Connection closing because of error: " << re.what() << endl;
-      run = false;
-    }
-  }
-  LOG(ERROR) << "Jumpclient shutdown";
-  close(routerFd);
-}
-
 int main(int argc, char **argv) {
   // Version string need to be set before GFLAGS parse arguments
   SetVersionString(string(ET_VERSION));
@@ -581,46 +401,6 @@ int main(int argc, char **argv) {
 
   if (FLAGS_port == 0) {
     FLAGS_port = 2022;
-  }
-
-  if (FLAGS_jump) {
-    string idpasskey = getIdpasskey();
-    string id = split(idpasskey, '/')[0];
-    string username = string(ssh_get_local_username());
-    // etserver with --jump cannot write to the default log file(root)
-    LogHandler::SetupLogFile(&defaultConf,
-                             "/tmp/etjump-" + username + "-" + id + ".log",
-                             maxlogsize);
-    // Reconfigure default logger to apply settings above
-    el::Loggers::reconfigureLogger("default", defaultConf);
-    // Install log rotation callback
-    el::Helpers::installPreRollOutCallback(LogHandler::rolloutHandler);
-
-    startJumpHostClient(idpasskey);
-
-    // Uninstall log rotation callback
-    el::Helpers::uninstallPreRollOutCallback();
-    return 0;
-  }
-
-  if (FLAGS_idpasskey.length() > 0 || FLAGS_idpasskeyfile.length() > 0) {
-    string idpasskey = getIdpasskey();
-    string id = split(idpasskey, '/')[0];
-    string username = string(ssh_get_local_username());
-    // etserver with --idpasskey cannot write to the default log file(root)
-    LogHandler::SetupLogFile(&defaultConf,
-                             "/tmp/etterminal-" + username + "-" + id + ".log",
-                             maxlogsize);
-    // Reconfigure default logger to apply settings above
-    el::Loggers::reconfigureLogger("default", defaultConf);
-    // Install log rotation callback
-    el::Helpers::installPreRollOutCallback(LogHandler::rolloutHandler);
-
-    startUserTerminal(idpasskey);
-
-    // Uninstall log rotation callback
-    el::Helpers::uninstallPreRollOutCallback();
-    return 0;
   }
 
   if (FLAGS_daemon) {
